Move daily to its own lib
